Fix Linux CI sign-conversion error and update CLAUDE.md for C++ SDK
Fix -Wsign-conversion error in integration_tests.cpp:394 where an int
loop variable was used to index a vector (requires size_t). This was
failing all 3 Linux CI jobs (gcc, clang-asan, linux-vcpkg).

Update CLAUDE.md from .NET SDK docs to C++ SDK: build commands for
Windows/Linux, debugging guides (Visual Studio, GDB, Docker, ASan),
known platform differences, and project structure.

+++ cpp/tests/integration/integration_tests.cpp
@@ -391,7 +391,7 @@ TEST_F(IntegrationTest, ConcurrentWorkflows) {
 
         // Wait for all results and verify.
         for (int i = 0; i < kNumWorkflows; ++i) {
-            auto result = run_task_sync(handles[i].get_result<int>());
+            auto result = run_task_sync(handles[static_cast<size_t>(i)].get_result<int>());
             EXPECT_EQ(result, i * 2) << "Workflow " << i << " returned wrong result";
         }
     });
